[LoginPage, Form-component, button-component] rename: rename form to -form-component and button
