+++ frontend/src/Components/Form/Form-component.module.scss
@@ -1,0 +1,41 @@
+@import '../../sass/settings.scss';
+
+.loginContainer__wrapper__loginFormContainer__header {
+  @include header-small;
+  &__alice {
+    @include aliceTextStarting;
+  }
+}
+
+.loginContainer__wrapper__loginFormContainer__form {
+    height: 100%;
+    @include d-flex-column;
+    gap: 2rem;
+    &__inputsWrapper {
+      @include d-flex-column;
+      justify-content: flex-start;
+      width: 100%;
+      &__label {
+        font-size: 1.3rem;
+        width: 100%;
+        padding-left: 1rem;
+      }
+      &__input {
+        width: 90%;
+        padding: 3% 10%;
+        border-radius: 0.3rem;
+        border: 2px solid #e4ccff;
+      }
+    }
+    &__buttonWrapper {
+      width: 2rem;
+    }
+    &__other {
+      @include d-flex-column;
+      &__content {
+        cursor: pointer;
+        text-decoration: underline;
+        color: #0275d8;
+      }
+    }
+  }
